feat: split audio_context from audio_context_config
