+++ packages/audioplayers_platform_interface/lib/src/audioplayers_platform.dart
@@ -4,7 +4,7 @@ import 'dart:async';
 // ignore: unnecessary_import
 import 'dart:typed_data';
 
-import 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
+import 'package:audioplayers_platform_interface/src/api/audio_context.dart';
 import 'package:audioplayers_platform_interface/src/api/player_event.dart';
 import 'package:audioplayers_platform_interface/src/api/player_mode.dart';
 import 'package:audioplayers_platform_interface/src/api/release_mode.dart';
